refactor: update alert-manager dir structure
