+++ apps/web/src/router/service-routes.ts
@@ -7,7 +7,7 @@ import { errorRoutes } from '@/router/error-routes';
 
 import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-constant';
 import administrationRoute from '@/services/administration/routes/routes';
-import alertManagerRoute from '@/services/alert-manager/routes';
+import alertManagerRoute from '@/services/alert-manager/routes/routes';
 import assetInventoryRoute from '@/services/asset-inventory/routes/routes';
 import authRoutes from '@/services/auth/routes/routes';
 import costExplorerRoute from '@/services/cost-explorer/routes/routes';
