+++ apps/web/src/services/alert-manager/alert-dashboard/modules/AlertStateWidget.vue
@@ -133,8 +133,8 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { red } from '@/styles/colors';
 
-import { ALERT_STATE } from '@/services/alert-manager/lib/config';
-import AlertListItem from '@/services/alert-manager/modules/AlertListItem.vue';
+import AlertListItem from '@/services/alert-manager/components/AlertListItem.vue';
+import { ALERT_STATE } from '@/services/alert-manager/constants/alert-constant';
 
 const TAB_STATE = Object.freeze({
     OPEN: 'OPEN',
