+++ apps/web/src/services/alert-manager/alert-dashboard/modules/project-search-widget/ProjectAlertListItem.vue
@@ -42,8 +42,8 @@ import type { UserReferenceMap } from '@/store/modules/reference/user/type';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { ALERT_STATE } from '@/services/alert-manager/lib/config';
-import AlertListItem from '@/services/alert-manager/modules/AlertListItem.vue';
+import AlertListItem from '@/services/alert-manager/components/AlertListItem.vue';
+import { ALERT_STATE } from '@/services/alert-manager/constants/alert-constant';
 import { PROJECT_ROUTE } from '@/services/project/routes/route-constant';
 
 export default {
