+++ apps/web/src/services/alert-manager/alert/alert-detail/modules/AlertStatusUpdate.vue
@@ -51,7 +51,7 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { useAlertPageStore } from '@/services/alert-manager/store/alert-page-store';
+import { useAlertPageStore } from '@/services/alert-manager/stores/alert-page-store';
 
 
 export default {
