+++ apps/web/src/services/alert-manager/alert/alert-detail/modules/alert-pushed-event/AlertPushedEvent.vue
@@ -116,7 +116,7 @@ import { copyAnyData } from '@/lib/helper/copy-helper';
 
 import AlertVerticalTimeline
     from '@/services/alert-manager/alert/alert-detail/modules/alert-pushed-event/modules/AlertVerticalTimeline.vue';
-import type { Event } from '@/services/alert-manager/type';
+import type { Event } from '@/services/alert-manager/types/alert-type';
 
 const PAGE_SIZE = 10;
