+++ apps/web/src/services/alert-manager/alert/alert-detail/modules/alert-summary/modules/AlertAssignModal.vue
@@ -52,8 +52,8 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useProxyValue } from '@/common/composables/proxy-state';
 
-import { useAlertPageStore } from '@/services/alert-manager/store/alert-page-store';
-import type { ProjectMember } from '@/services/alert-manager/type';
+import { useAlertPageStore } from '@/services/alert-manager/stores/alert-page-store';
+import type { ProjectMember } from '@/services/alert-manager/types/alert-type';
 
 
 export default {
