+++ apps/web/src/services/alert-manager/alert/modules/AlertAcknowledgeModal.vue
@@ -37,8 +37,8 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useProxyValue } from '@/common/composables/proxy-state';
 
-import { ALERT_STATE } from '@/services/alert-manager/lib/config';
-import type { AlertStateUpdateParams } from '@/services/alert-manager/type';
+import { ALERT_STATE } from '@/services/alert-manager/constants/alert-constant';
+import type { AlertStateUpdateParams } from '@/services/alert-manager/types/alert-type';
 
 export default {
     name: 'AlertAcknowledgeModalModal',
