+++ apps/web/src/services/alert-manager/components/AlertListItem.vue
@@ -74,9 +74,9 @@ import { referenceRouter } from '@/lib/reference/referenceRouter';
 import { red } from '@/styles/colors';
 
 import { useAlertStateI18n } from '@/services/alert-manager/composables/alert-state-i18n';
-import { ALERT_MANAGER_ROUTE } from '@/services/alert-manager/route-config';
+import { ALERT_MANAGER_ROUTE } from '@/services/alert-manager/routes/route-constant';
 
-import { ALERT_STATE } from '../lib/config';
+import { ALERT_STATE } from '../constants/alert-constant';
 
 const ALERT_URGENCY = Object.freeze({
     HIGH: 'HIGH',
