+++ apps/web/src/services/alert-manager/helpers/alert-badge-helper.ts
@@ -1,4 +1,5 @@
-import { ALERT_STATE, SCOPE } from '@/services/alert-manager/lib/config';
+import { ALERT_STATE, SCOPE } from '@/services/alert-manager/constants/alert-constant';
+
 
 export const alertStateBadgeStyleTypeFormatter = (alertState) => {
     let style = '';
