+++ apps/web/src/services/alert-manager/types/escalation-policy-type.ts
@@ -1,0 +1,2 @@
+
+export type EscalationPolicyAction = 'CREATE' | 'UPDATE' | 'DELETE' | 'DEFAULT';
