+++ apps/web/src/services/project/components/ProjectSummaryAlertWidget.vue
@@ -79,9 +79,9 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { alert, secondary } from '@/styles/colors';
 
-import { ALERT_STATE } from '@/services/alert-manager/lib/config';
-import AlertListItem from '@/services/alert-manager/modules/AlertListItem.vue';
-import type { AlertListPageUrlQuery } from '@/services/alert-manager/type';
+import AlertListItem from '@/services/alert-manager/components/AlertListItem.vue';
+import { ALERT_STATE } from '@/services/alert-manager/constants/alert-constant';
+import type { AlertListPageUrlQuery } from '@/services/alert-manager/types/alert-type';
 import { PROJECT_ROUTE } from '@/services/project/routes/route-constant';
 
 const ASSIGNED_STATE = Object.freeze({
