+++ apps/web/src/services/project/pages/ProjectAlertListPage.vue
@@ -9,8 +9,8 @@ import { QueryHelper } from '@cloudforet/core-lib/query';
 import { useManagePermissionState } from '@/common/composables/page-manage-permission';
 
 import AlertDataTable from '@/services/alert-manager/alert/modules/AlertDataTable.vue';
-import { ALERT_STATE_FILTER, ALERT_URGENCY, ASSIGNED_STATE } from '@/services/alert-manager/lib/config';
-import type { AlertListPageUrlQuery, AlertListTableFilters } from '@/services/alert-manager/type';
+import { ALERT_STATE_FILTER, ALERT_URGENCY, ASSIGNED_STATE } from '@/services/alert-manager/constants/alert-constant';
+import type { AlertListPageUrlQuery, AlertListTableFilters } from '@/services/alert-manager/types/alert-type';
 import { useProjectDetailPageStore } from '@/services/project/stores/project-detail-page-store';
 
 interface Props {
